Port export-just-namespace
