+++ packages/eslint-plugin/test/__file_snapshots__/types/export-just-namespace/bad1.d.ts.lint
@@ -1,0 +1,11 @@
+types/export-just-namespace/bad1.d.ts
+  1:1  error  Instead of `export =`-ing a namespace, use the body of the namespace as the module body  @definitelytyped/export-just-namespace
+
+✖ 1 problem (1 error, 0 warnings)
+
+==== types/export-just-namespace/bad1.d.ts ====
+
+    export = Stuff;
+    ~~~~~~~~~~~~~~~
+!!! @definitelytyped/export-just-namespace: Instead of `export =`-ing a namespace, use the body of the namespace as the module body.
+    namespace Stuff {}
