+++ packages/eslint-plugin/test/__file_snapshots__/types/export-just-namespace/export-just-namespace-tests.ts.lint
@@ -1,0 +1,10 @@
+types/export-just-namespace/export-just-namespace-tests.ts
+  1:1  error  File has no content  @definitelytyped/no-useless-files
+
+✖ 1 problem (1 error, 0 warnings)
+
+==== types/export-just-namespace/export-just-namespace-tests.ts ====
+
+    
+    ~
+!!! @definitelytyped/no-useless-files: File has no content.
