+++ packages/eslint-plugin/test/fixtures/types/export-just-namespace/bad1.d.ts
@@ -1,0 +1,2 @@
+export = Stuff;
+namespace Stuff {}
