+++ packages/eslint-plugin/test/fixtures/types/export-just-namespace/bad2.d.ts
@@ -1,0 +1,2 @@
+namespace Stuff {}
+export = Stuff;
