+++ packages/eslint-plugin/test/fixtures/types/export-just-namespace/bad3.d.ts
@@ -1,0 +1,3 @@
+namespace Stuff {}
+const other = "code";
+export = Stuff;
