+++ packages/eslint-plugin/test/fixtures/types/export-just-namespace/good1.d.ts
@@ -1,0 +1,1 @@
+export const value = 3;
