+++ packages/eslint-plugin/test/fixtures/types/export-just-namespace/good2.d.ts
@@ -1,0 +1,1 @@
+export default class Hello {}
