+++ packages/eslint-plugin/test/fixtures/types/export-just-namespace/good3.d.ts
@@ -1,0 +1,1 @@
+import * as fs from "fs";
