+++ packages/eslint-plugin/test/fixtures/types/export-just-namespace/good4.d.ts
@@ -1,0 +1,1 @@
+const value = 123;
