+++ packages/eslint-plugin/test/fixtures/types/export-just-namespace/good5.d.ts
@@ -1,0 +1,1 @@
+export = stuff;
