+++ packages/eslint-plugin/test/fixtures/types/export-just-namespace/good6.d.ts
@@ -1,0 +1,1 @@
+export = createStuff();
