+++ packages/eslint-plugin/test/fixtures/types/export-just-namespace/good7.d.ts
@@ -1,0 +1,3 @@
+class Stuff {}
+namespace Stuff {}
+export = Stuff;
