+++ packages/eslint-plugin/test/fixtures/types/export-just-namespace/good8.d.ts
@@ -1,0 +1,3 @@
+export = First
+namespace First {}
+declare function First()
