+++ packages/eslint-plugin/test/fixtures/types/export-just-namespace/good9.d.ts
@@ -1,0 +1,3 @@
+declare namespace Second {}
+export = Second
+declare function Second<U, S>(s: U): S
