+++ packages/eslint-plugin/test/fixtures/types/export-just-namespace/tsconfig.json
@@ -1,0 +1,31 @@
+{
+    "compilerOptions": {
+        "module": "commonjs",
+        "lib": [
+            "es6",
+            "dom"
+        ],
+        "noImplicitAny": true,
+        "noImplicitThis": true,
+        "strictFunctionTypes": true,
+        "strictNullChecks": true,
+        "types": [],
+        "noEmit": true,
+        "forceConsistentCasingInFileNames": true
+    },
+    "files": [
+        "good1.d.ts",
+        "good2.d.ts",
+        "good3.d.ts",
+        "good4.d.ts",
+        "good5.d.ts",
+        "good6.d.ts",
+        "good7.d.ts",
+        "good8.d.ts",
+        "good9.d.ts",
+        "bad1.d.ts",
+        "bad2.d.ts",
+        "bad3.d.ts",
+        "export-just-namespace-tests.ts"
+    ]
+}
